fix(#1234/#1235): -c=native long long spelling + i++ discard-context
#1234: CCCC's type system has no distinct `long long` kind -- `long` and
`long long` both fold to the `ty_long` singleton in parse_types.c's
declspec, and serialize_type.c's TY_LONG arm always spelled `long`. A
header compiled as ordinary C alongside the generated TU (its `#include`
replayed verbatim) that itself writes `long long` then collided with the
re-emitted `long` prototype -- "conflicting types". Add a
serialization-only Type.is_long_long spelling bit plus ty_llong/ty_ullong
singletons (same {TY_LONG, 8, 8}); the `long long` declspec cases select
them and serialize_type reproduces the spelling. compat_body /
same_type_or_origin deliberately do not compare the bit -- `long` and
`long long` stay one type for rank, _Generic, and compatibility.
examples/ccccl's runtime is restored to `long long` (was the workaround)
and is now the real-world regression test.

#1235: new_inc_dec() lowers `A++` to `(typeof A)((A += 1) + -1)`; the
`+ -1` reconstructs the pre-increment value and is dead in a
value-discarding position, drawing -Wunused-value on every
`for (i = 0; i < n; i++)` in -c=native output. New serialize_discard_expr()
peels the is_inc_dec_result cast, the compensating ND_ADD, and any
usual_arith_conv promotion cast down to the RMW comma; wired into
serialize_stmt.c's ND_EXPR_STMT arm and the for-update clause. A
value-position `j = i++` keeps the full form.

Tests: tests/test_native_long_long_1234.c (+ companion header),
tests/test_native_inc_dec_discard_1235.c, and smoke cases 151/152 in
tools/comptime_native_smoke.py (case 152 also compiles the -m output
under -Wunused-value -Werror). Full VM + native (clang, gcc-16) suites
green. man/NATIVE.md's divergence table drops the #1234 row.

+++ examples/ccccl/runtime/ccccl_rt.c
@@ -79,7 +79,7 @@ LObj *ccccl_intern(const char *name) {
     }
 }
 
-LObj *ccccl_int(long v) {
+LObj *ccccl_int(long long v) {
     LObj *n;
     if (g_int_next >= CCCCL_RT_MAX_INTS)
         ccccl_rt_fatal("int arena exhausted");
@@ -122,7 +122,7 @@ LObj *ccccl_eq(LObj *a, LObj *b) {
     return ccccl_nil;
 }
 
-static long ccccl_as_int(LObj *x, const char *who) {
+static long long ccccl_as_int(LObj *x, const char *who) {
     if (x->tag != CCCCL_INT)
         ccccl_rt_fatal2(who, ": not a number");
     return x->as.ival;
@@ -138,13 +138,13 @@ LObj *ccccl_mul(LObj *a, LObj *b) {
     return ccccl_int(ccccl_as_int(a, "*") * ccccl_as_int(b, "*"));
 }
 LObj *ccccl_div(LObj *a, LObj *b) {
-    long bv = ccccl_as_int(b, "/");
+    long long bv = ccccl_as_int(b, "/");
     if (bv == 0)
         ccccl_rt_fatal("/: division by zero");
     return ccccl_int(ccccl_as_int(a, "/") / bv);
 }
 LObj *ccccl_mod(LObj *a, LObj *b) {
-    long bv = ccccl_as_int(b, "mod");
+    long long bv = ccccl_as_int(b, "mod");
     if (bv == 0)
         ccccl_rt_fatal("mod: division by zero");
     return ccccl_int(ccccl_as_int(a, "mod") % bv);
@@ -191,7 +191,7 @@ LObj *ccccl_closure_self(CccclFn fn, LObj *captures, int self_slot) {
 
 void ccccl_print(LObj *x, FILE *out) {
     if (x->tag == CCCCL_INT) {
-        fprintf(out, "%ld", x->as.ival);
+        fprintf(out, "%lld", x->as.ival);
         return;
     }
     if (x->tag == CCCCL_CLOSURE) {
